requirements.txt
Django==1.6
South==0.8.4
django-autoslug==1.7.1
django-braces==1.2.2
six==1.4.1
wsgiref==0.1.2
